Handle encryption keys only internally (closes #61)
The key is unique per asarmor installation and should be the exact same from both the JS and C/C++ side. To prevent user error we should not expose the `key` property to the user in the library API or CLI. Instead, we always use the randomly generated key. If a user wishes to rotate the key, they can do so by manually reinstalling `asarmor` or by deleting the `key.txt` file and regenerating it using a script.